changes in mnode

+++ DistrSim/readgraph.cpp
@@ -33,8 +33,8 @@ readGraph::readGraph(MNode *n)
         QString str(line);
         QStringList strList = str.split(" ");
 
-        exampleNode *sn = new exampleNode(strList[0]);
-        exampleNode *dn = new exampleNode(strList[1]);
+        exampleNode *sn = new exampleNode(strList[0].toLong());
+        exampleNode *dn = new exampleNode(strList[1].toLong());
 
         exampleEdge *ee = new exampleEdge(sn, dn);
